Support filtering model
Introduced a new component FilterSet, that can be used to define
a list of the filters and sets a workflow to them.

The filters itself lives in the domain model and UI component
wires them together (it may be a Page component or other compopent
that is not abstract over a filtes, in our case it's a BookFilter).

Unfortunately the configuration of the supported filters looks not
as clean as it should, we have to have some boilerplate to pass
handlers inside.